Refactor: Isolate XY Plot UI into xy_plot_app.py
To improve maintainability and clean up interface.py, I've moved all the UI components and logic for the XY Plot feature into a dedicated modules/xy_plot_app.py file.

This should make it easier for the team to work on this feature without causing conflicts in the main interface file.
